Move ForkAwesome to common
