+++ src/components/common/icon-button/icon-button.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Button, ButtonProps } from 'react-bootstrap'
-import { ForkAwesomeIcon, IconName } from '../../fork-awesome/fork-awesome-icon'
+import { ForkAwesomeIcon, IconName } from '../fork-awesome/fork-awesome-icon'
 import './icon-button.scss'
 
 export interface SocialButtonProps extends ButtonProps {
