+++ src/components/landing/layout/navigation/new-guest-note-button.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import { LinkContainer } from 'react-router-bootstrap'
 import { Button } from 'react-bootstrap'
 import { Trans, useTranslation } from 'react-i18next'
-import { ForkAwesomeIcon } from '../../../../fork-awesome/fork-awesome-icon'
+import { LinkContainer } from 'react-router-bootstrap'
+import { ForkAwesomeIcon } from '../../../common/fork-awesome/fork-awesome-icon'
 
 export const NewGuestNoteButton: React.FC = () => {
   const { t } = useTranslation()
