+++ src/components/landing/pages/history/history-card/history-card.tsx
@@ -1,8 +1,8 @@
 import moment from 'moment'
 import React from 'react'
 import { Badge, Card } from 'react-bootstrap'
-import { ForkAwesomeIcon } from '../../../../../fork-awesome/fork-awesome-icon'
 import { formatHistoryDate } from '../../../../../utils/historyUtils'
+import { ForkAwesomeIcon } from '../../../../common/fork-awesome/fork-awesome-icon'
 import { CloseButton } from '../common/close-button'
 import { PinButton } from '../common/pin-button'
 import { SyncStatus } from '../common/sync-status'
