+++ src/components/landing/pages/history/history-toolbar/export-history-button.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Button } from 'react-bootstrap'
 import { useTranslation } from 'react-i18next'
-import { ForkAwesomeIcon } from '../../../../../fork-awesome/fork-awesome-icon'
+import { ForkAwesomeIcon } from '../../../../common/fork-awesome/fork-awesome-icon'
 
 export interface ExportHistoryButtonProps {
   onExportHistory: () => void
